requirements.txt
chess==1.11.2
keyboard==0.13.5
pillow==11.1.0
